Tests/LibWeb/Text/input/css-has-invalidation/has-sibling-scan-insights.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style id="test-style">
    .anchor:has(.match) { color: red; }
</style>
<div id="outer">
    <div id="anchor" class="anchor">
        <div><span id="target">target</span></div>
    </div>
    <div>unrelated sibling</div>
</div>
<script>
    function printSiblingChecks(label) {
        let counters = internals.getStyleInvalidationCounters();
        println(`[${label}]`);
        println(`  hasAncestorWalkInvocations: ${counters.hasAncestorWalkInvocations}`);
        println(`  hasAncestorWalkVisits: ${counters.hasAncestorWalkVisits}`);
        println(`  hasAncestorSiblingElementChecks: ${counters.hasAncestorSiblingElementChecks}`);
        println(`  hasMatchInvocations: ${counters.hasMatchInvocations}`);
        println(`  hasResultCacheHits: ${counters.hasResultCacheHits}`);
        println(`  hasResultCacheMisses: ${counters.hasResultCacheMisses}`);
        println(`  styleInvalidations: ${counters.styleInvalidations}`);
    }

    test(() => {
        let anchor = document.getElementById("anchor");
        let target = document.getElementById("target");
        let style = document.getElementById("test-style");

        getComputedStyle(anchor).color;
        internals.resetStyleInvalidationCounters();

        target.classList.add("match");
        getComputedStyle(anchor).color;
        printSiblingChecks("descendant-only :has() sheet");

        target.classList.remove("match");
        style.textContent += "\n.unused-anchor:has(+ .unused-match) { color: blue; }";
        getComputedStyle(anchor).color;
        internals.resetStyleInvalidationCounters();

        target.classList.add("match");
        getComputedStyle(anchor).color;
        printSiblingChecks("sheet with sibling-sensitive :has()");
    });
</script>
