Tests/LibWeb/Text/expected/css-has-invalidation/has-sibling-scan-insights.txt

[descendant-only :has() sheet]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 6
  hasAncestorSiblingElementChecks: 0
  hasMatchInvocations: 1
  hasResultCacheHits: 0
  hasResultCacheMisses: 1
  styleInvalidations: 1
[sheet with sibling-sensitive :has()]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 6
  hasAncestorSiblingElementChecks: 10
  hasMatchInvocations: 1
  hasResultCacheHits: 0
  hasResultCacheMisses: 1
  styleInvalidations: 1
